# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VoluntaRed API - Plataforma de voluntariado juvenil con sistema de matching
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBodyUserType =
  (typeof RegisterBodyUserType)[keyof typeof RegisterBodyUserType];

export const RegisterBodyUserType = {
  volunteer: "volunteer",
  organization: "organization",
} as const;

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  userType: RegisterBodyUserType;
  /** @nullable */
  organizationName?: string | null;
  /** @nullable */
  organizationDescription?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserUserType = (typeof UserUserType)[keyof typeof UserUserType];

export const UserUserType = {
  volunteer: "volunteer",
  organization: "organization",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  userType: UserUserType;
  /** @nullable */
  organizationName?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  userType: string;
  /** @nullable */
  organizationName?: string | null;
  /** @nullable */
  organizationDescription?: string | null;
  /** @nullable */
  bio?: string | null;
  skills: string[];
  interests: string[];
  accessibilityNeeds: string[];
  /** @nullable */
  state?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  age?: number | null;
  createdAt: string;
}

export interface UpdateProfileBody {
  name?: string;
  bio?: string;
  skills?: string[];
  interests?: string[];
  accessibilityNeeds?: string[];
  state?: string;
  city?: string;
  age?: number;
  organizationName?: string;
  organizationDescription?: string;
}

export type OpportunityEffortLevel =
  (typeof OpportunityEffortLevel)[keyof typeof OpportunityEffortLevel];

export const OpportunityEffortLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Opportunity {
  id: number;
  title: string;
  description: string;
  organizationId: number;
  organizationName: string;
  category: string;
  skills: string[];
  interests: string[];
  accessibilityFeatures: string[];
  effortLevel: OpportunityEffortLevel;
  location: string;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  spotsAvailable?: number | null;
  isActive: boolean;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface OpportunityDetail {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  requirements?: string | null;
  organizationId: number;
  organizationName: string;
  /** @nullable */
  organizationDescription?: string | null;
  category: string;
  skills: string[];
  interests: string[];
  accessibilityFeatures: string[];
  effortLevel: string;
  location: string;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  spotsAvailable?: number | null;
  applicantsCount: number;
  isActive: boolean;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export type CreateOpportunityBodyEffortLevel =
  (typeof CreateOpportunityBodyEffortLevel)[keyof typeof CreateOpportunityBodyEffortLevel];

export const CreateOpportunityBodyEffortLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateOpportunityBody {
  title: string;
  description: string;
  requirements?: string;
  category: string;
  skills?: string[];
  interests?: string[];
  accessibilityFeatures?: string[];
  effortLevel: CreateOpportunityBodyEffortLevel;
  location: string;
  state?: string;
  startDate?: string;
  endDate?: string;
  spotsAvailable?: number;
}

export interface UpdateOpportunityBody {
  title?: string;
  description?: string;
  requirements?: string;
  category?: string;
  skills?: string[];
  interests?: string[];
  accessibilityFeatures?: string[];
  effortLevel?: string;
  location?: string;
  state?: string;
  startDate?: string;
  endDate?: string;
  spotsAvailable?: number;
  isActive?: boolean;
}

export interface ApplyBody {
  message: string;
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface Application {
  id: number;
  opportunityId: number;
  opportunityTitle: string;
  userId: number;
  userName: string;
  message: string;
  status: ApplicationStatus;
  createdAt: string;
}

export interface MatchedOpportunity {
  opportunity: Opportunity;
  matchScore: number;
  matchReasons: string[];
}

export interface MatchScore {
  opportunityId: number;
  totalScore: number;
  skillsScore: number;
  interestsScore: number;
  accessibilityScore: number;
  locationScore: number;
  reasons: string[];
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface StatsSummary {
  totalOpportunities: number;
  totalVolunteers: number;
  totalOrganizations: number;
  totalApplications: number;
  categoryCounts: CategoryCount[];
}

export interface Category {
  id: string;
  name: string;
  icon: string;
}

export type ListOpportunitiesParams = {
  category?: string;
  accessibility?: string;
  effort?: string;
  state?: string;
  search?: string;
  skills?: string;
};