# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VoluntaRed API - Plataforma de voluntariado juvenil con sistema de matching
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /matching/score/{opportunityId}:
    get:
      operationId: getMatchScore
      tags: [matching]
      summary: Get match score for a specific opportunity
      parameters:
        - name: opportunityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match score and breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchScore"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        userType:
          type: string
          enum: [volunteer, organization]
        organizationName:
          type: ["string", "null"]
        organizationDescription:
          type: ["string", "null"]
      required:
        - email
        - password
        - name
        - userType
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        userType:
          type: string
          enum: [volunteer, organization]
        organizationName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - userType
        - createdAt
...
    MatchScore:
      type: object
      properties:
        opportunityId:
          type: integer
        totalScore:
          type: integer
        skillsScore:
          type: integer
        interestsScore:
          type: integer
        accessibilityScore:
          type: integer
        locationScore:
          type: integer
        reasons:
          type: array
          items:
            type: string
      required:
        - opportunityId
        - totalScore
        - skillsScore
        - interestsScore
        - accessibilityScore
        - locationScore
        - reasons
...