# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VoluntaRed API - Plataforma de voluntariado juvenil con sistema de matching
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterBodyUserType } from "./registerBodyUserType";

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  userType: RegisterBodyUserType;
  /** @nullable */
  organizationName?: string | null;
  /** @nullable */
  organizationDescription?: string | null;
}