# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VoluntaRed API - Plataforma de voluntariado juvenil con sistema de matching
 * OpenAPI spec version: 0.1.0
 */
import type { UserUserType } from "./userUserType";

export interface User {
  id: number;
  email: string;
  name: string;
  userType: UserUserType;
  /** @nullable */
  organizationName?: string | null;
  createdAt: string;
}